api/requirements.txt
# Core dependencies
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy[asyncio]==2.0.21
pydantic==2.4.2
pydantic-settings==2.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncpg==0.28.0
alembic==1.12.0

# Testing dependencies
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1
